+++ deltafi-core/src/main/java/org/deltafi/core/types/TimedDeletePolicy.java
@@ -30,7 +30,6 @@ public class TimedDeletePolicy implements DeletePolicy {
     private String id;
     private String name;
     private boolean enabled;
-    private boolean locked;
     private String flow;
     private String afterCreate;
     private String afterComplete;